+++ src/ll/misc.py
@@ -15,7 +15,7 @@ LivingLogic modules and packages.
 """
 
 
-import sys, os, os.path, platform, types, datetime, collections, io, gzip as gzip_, argparse, functools, signal, contextlib, subprocess
+import sys, os, os.path, platform, types, datetime, collections, io, gzip as gzip_, argparse, functools, signal, contextlib, subprocess, enum
 from collections import abc
 
 from ll import color
@@ -99,7 +99,7 @@ def item(iterable, index, default=None):
 	exhausted up to the ``index``'th item. If the iterator doesn't produce
 	that many items ``default`` will be returned.
 
-	``index`` may also be an iterable of indexes, in which case :meth:`item`
+	``index`` may also be an iterable of indexes, in which case :func:`!item`
 	will be applied recursively, i.e. ``item(["foo", "bar"], (1, -1))`` returns
 	``'r'``.
 	"""
@@ -257,6 +257,29 @@ def withdoc(doc):
 	return wrapper
 
 
+class EnumMeta(enum.Enum.__class__):
+	def __repr__(self):
+		return f"<enum {self.__module__}.{self.__qualname__}>"
+
+
+class _EnumRepr:
+	def __repr__(self):
+		return f"<{self.__class__.__module__}.{self.__class__.__qualname__}({self.value!r})>"
+
+class Enum(_EnumRepr, enum.Enum, metaclass=EnumMeta):
+	"""
+	Subclass of :class:`enum.Enum` where class and instance :func:`repr` output
+	include the module and fully qualified class name.
+	"""
+
+
+class IntEnum(_EnumRepr, enum.IntEnum, metaclass=EnumMeta):
+	"""
+	Subclass of :class:`enum.IntEnum` where class and instance :func:`repr`
+	output includes the module and fully qualified class name.
+	"""
+
+
 class _propclass_Meta(type):
 	def __new__(cls, name, bases, dict):
 		if bases == (property,):
@@ -280,9 +303,9 @@ class _propclass_Meta(type):
 
 class propclass(property, metaclass=_propclass_Meta):
 	'''
-	:class:`propclass` provides an alternate way to define properties.
+	:class:`!propclass` provides an alternate way to define properties.
 
-	Subclassing :class:`propclass` and defining methods :meth:`__get__`,
+	Subclassing :class:`!propclass` and defining methods :meth:`__get__`,
 	:meth:`__set__` and :meth:`__delete__` will automatically generate the
 	appropriate property::
 
@@ -351,7 +374,7 @@ def exception_chain(exc):
 
 class Pool:
 	"""
-	A :class:`Pool` object can be used as an inheritable alternative to modules.
+	A :class:`!Pool` object can be used as an inheritable alternative to modules.
 	The attributes of a module can be put into a pool and each pool can have
 	base pools where lookup continues if an attribute can't be found.
 	"""
@@ -365,7 +388,7 @@ class Pool:
 	def register(self, object):
 		r"""
 		Register ``object`` in the pool. ``object`` can be a module, a
-		dictionary or a :class:`Pool` objects (with registers the pool as a base
+		dictionary or a :class:`!Pool` objects (with registers the pool as a base
 		pool). If ``object`` is a module and has an attribute :attr:`__bases__`
 		(being a sequence of other modules) this attribute will be used to
 		initialize ``self``\s base pool.
@@ -432,7 +455,7 @@ def iterone(item):
 
 class Iterator:
 	"""
-	:class:`Iterator` adds :meth:`__getitem__` support to an iterator. This is
+	:class:`!Iterator` adds :meth:`__getitem__` support to an iterator. This is
 	done by calling :func:`item` internally.
 	"""
 	__slots__ = ("iterator", )
@@ -473,7 +496,7 @@ class Iterator:
 
 class Queue:
 	"""
-	:class:`Queue` provides FIFO queues: The method :meth:`write` writes to the
+	:class:`!Queue` provides FIFO queues: The method :meth:`write` writes to the
 	queue and the method :meth:`read` read from the other end of the queue and
 	remove the characters read.
 	"""
@@ -518,7 +541,7 @@ class Const:
 
 class FlagAction(argparse.Action):
 	"""
-	:class:`FlagAction` can be use with :mod:`argparse` for options that
+	:class:`!FlagAction` can be use with :mod:`argparse` for options that
 	represent flags. An options can have a value like ``yes`` or ``no`` for the
 	correspending boolean value, or if the value is omitted it is the inverted
 	default value (i.e. specifying the option toggles it).
@@ -697,7 +720,7 @@ def javaexpr(obj):
 
 class SysInfo:
 	"""
-	A :class:`SysInfo` object contains information about the host, user, python
+	A :class:`!SysInfo` object contains information about the host, user, python
 	version and script. Available attributes are ``host_name``, ``host_fqdn``,
 	``host_ip``, ``host_sysname``, ``host_nodename``, ``host_release``,
 	``host_version``, ``host_machine``, ``user_name``, ``user_uid``, ``user_gid``,
@@ -705,7 +728,7 @@ class SysInfo:
 	``python_version``, ``pid``, ``script_name``, ``short_script_name`` and
 	``script_url``.
 
-	:class:`SysInfo` object also support a mimimal dictionary interface (i.e.
+	:class:`!SysInfo` object also support a mimimal dictionary interface (i.e.
 	:meth:`__getitem__` and :meth:`__iter__`).
 
 	One module global instance named ``sysinfo`` is created at module import
@@ -908,7 +931,7 @@ sysinfo = SysInfo()
 
 class monthdelta:
 	"""
-	:class:`monthdelta` objects can be used to add months/years to a
+	:class:`!monthdelta` objects can be used to add months/years to a
 	:class:`datetime.datetime` or :class:`datetime.date` object. If the resulting
 	day falls out of the range of valid days for the target month, the last day
 	for the target month will be used instead::
@@ -1049,7 +1072,7 @@ def timeout(seconds):
 	"""
 	A context manager that limits the runtime of the wrapped code.
 
-	This doesn't work with threads and only on UNIX.
+	As this uses :mod:`signal`, this won't with threads and only on UNIX.
 	"""
 
 	def _timeouthandler(signum, frame):
@@ -1091,7 +1114,7 @@ def prettycsv(rows, padding="   "):
 	``rows`` must be a list of lists of strings (e.g. as produced by the
 	:mod:`csv` module). ``padding`` is the padding between columns.
 
-	:func:`prettycsv` is a generator.
+	:func:`!prettycsv` is a generator.
 	"""
 
 	def width(row, i):