+++ docs/NEWS.rst
@@ -14,6 +14,17 @@ Changes in 5.50 (released 07/??/2019)
 *	There's a new option ``-a``/``--ascii`` for running PySQL scripts: With this
 	PySQL will not use unicode characters for drawing fancy boxes.
 
+*	Fixed a bug in the filename handling of PySQL, so that showing source in
+	stack traces work again: as we're changing directories now, using relative
+	paths no longer worked.
+
+*	Fixed logic for showing line numbers for locations in PySQL scripts.
+
+*	Added classes :class:`ll.misc.Enum` and :class:`ll.misc.IntEnum` that are
+	subclasses of :class:`enum.Enum` and :class:`enum.IntEnum`, but show the
+	module and fully qualified class name in the :func:`repr` output for classes
+	and instances.
+
 
 Changes in 5.49 (released 07/04/2019)
 -------------------------------------