+++ docs/NEWS.rst
@@ -24,6 +24,8 @@ Changes in 5.50 (released 07/??/2019)
 
 *	Fixed logic for showing line numbers for locations in PySQL scripts.
 
+*	Fixed a bug in the PySQL command ``resetsequence``.
+
 *	Added classes :class:`ll.misc.Enum` and :class:`ll.misc.IntEnum` that are
 	subclasses of :class:`enum.Enum` and :class:`enum.IntEnum`, but show the
 	module and fully qualified class name in the :func:`repr` output for classes